fix(coldflow): Pressure control on tanks first, then switch to venturis
During coldflow the venturi pressures were to high due to the pressure
control target sensors being on venturis from beginning. This lead to
a high tank pressure before opening main valves due to always low pressure on venturis.
The pressure control closed the pressurization as expected, however, the high pressure in the
tank still led to too high venturi pressures.

To avoid this, the pressure control regulates the tank pressure until
the main valves are opened and only then switches to venturi pressure.